Réécrire volume settings en français
